re-add handling of import tags as deprecated
Future versions of perl are intending to make it an error to pass
arguments to an undefined import method. This module was relying on the
arguments to be ignored.

Re-add an import method that warns about the no longer needed tags,
adjust the tests to account for that, and remove the example using :pg
from the docs.